moved templated implementations into a detail directory
